changes from pull request comments
