+++ recordlinkage/contrib/index/__init__.py
@@ -24,3 +24,4 @@
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
+from .neighbourhoodblock import NeighbourhoodBlock
